requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.21.0=pypi_0
aiofiles=23.1.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.5.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=pypi_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py39h9ce1e76_2
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.1.2=pypi_0
fastapi=0.100.0=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.1=pypi_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.41.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fschat=0.2.3=pypi_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gradio=3.23.0=pypi_0
grpcio=1.51.3=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=py39h06a4308_0
importlib-resources=6.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.18.4=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.0.149=0
libcurand=10.3.3.53=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
linkify-it-py=2.0.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.2.0=pypi_0
markdown2=2.4.9=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.7.2=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
msgpack=1.0.5=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
ninja=1.11.1=pypi_0
numpy=1.25.0=py39h5f9d8c6_0
numpy-base=1.25.0=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.9=h7f8727e_0
orjson=3.9.2=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.1.2=py39h06a4308_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.11=pypi_0
pydub=0.25.1=pypi_0
pygments=2.15.1=pypi_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=pypi_0
pyre-extensions=0.0.29=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytorch=2.0.1=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
ray=2.5.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.0=pypi_0
regex=2023.6.3=pypi_0
requests=2.29.0=py39h06a4308_0
rich=13.4.2=pypi_0
rpds-py=0.9.2=pypi_0
safetensors=0.3.1=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.28.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=67.8.0=py39h06a4308_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.27.0=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.2=py39_cu118
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu118
tqdm=4.65.0=pypi_0
transformers=4.31.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.6.3=py39h06a4308_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
urllib3=1.26.16=py39h06a4308_0
uvicorn=0.23.1=pypi_0
vllm=0.1.2=pypi_0
wandb=0.15.5=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.6=pypi_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py39h06a4308_0
xformers=0.0.20=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
